Added Many Features
Added the Following:
Texture Coordinates (Showing Part of a Texture)
Renderable Rotations (Rendering the Texture Rotated)
Increased Rendering Efficiency
Text File Clean-Up
Text with Multiple Fonts and/or Colors
Newlines in Text
DynamicBox (A box with a frame that has a dynamic size and only uses one Texture)
TextBox (A DynamicBox that fits snuggly around a Text)

+++ src/handlers/Cursor.java
@@ -12,6 +12,6 @@ public class Cursor extends GLFWCursorPosCallback {
         yPos = y;
     }
 
-    public static double x() {return xPos;}
-    public static double y() {return yPos;}
+    public static float x() {return (float) xPos;}
+    public static float y() {return (float) yPos;}
 }
